Access params for buffer size and hopefully finish double box detection
I made it so it accesses the params file to get the value for the buffer it will give when determining if it is a 2 box senario. I think it might be good to allow modification of + buffer or - buffer. I also played around with the array to fix the issue where it would only get the center coordinates of one of the boxes. Now it should get the center for both. (please review)